Move functions in its respected cubit
