+++ app/lib/actions/zoom.dart
@@ -36,9 +36,13 @@ class ZoomAction extends Action<ZoomIntent> {
       (viewport.height ?? 0) / 2,
     );
     final transformCubit = cubit.transformCubit;
-    cubit.size(
+    cubit.transformCubit.sizeConstrained(
       transformCubit.state.size + (intent.reverse ? -0.1 : 0.1),
-      center,
+      cursor: center,
+      settingsCubit: cubit.settingsCubit,
+      rendererCubit: cubit.rendererCubit,
+      inputCubit: cubit.inputCubit,
+      viewCubit: cubit.viewCubit,
     );
   }
 }
