+++ app/lib/handlers/shape.dart
@@ -42,7 +42,9 @@ class ShapeHandler extends PastingHandler<ShapeTool> with ColoredHandler {
         property: data.property.copyWith(
           strokeWidth:
               data.property.strokeWidth /
-              (data.zoomDependent ? cubit.rendererCubit.state.cameraViewport.scale : 1),
+              (data.zoomDependent
+                  ? cubit.rendererCubit.state.cameraViewport.scale
+                  : 1),
         ),
         collection: collection,
       ),
